+++ NEWS.md
@@ -1,7 +1,9 @@
-ctmm 0.5.1 (2018-05-17)
+ctmm 0.5.1 (2018-05-28)
 ================
 
-  * Units of measurement down to microns and microseconds now supported.
+  * Units of measurement down to microns and microseconds now supported
+  
+  * fixed bug in as.telemetry with projection argument
   
   * fixed bug in new variogram error calculations (v0.5.0) used for plotting